refactor(head-meta): move HeadMeta model to its own file
